# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/order": {
            "get": {
                "tags": [
                    "order"
                ],
                "summary": "Returns all the orders in list.",
                "description": "Returns all the orders saved in database in list.",
                "operationId": "get orders",
                "responses": {
                    "200": {
                        "description": "Successfully got the order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "datas": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nothing found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not found"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "order"
                ],
                "summary": "Store order from post form",
                "description": "Store a new order from form in database, using post method",
                "operationId": "Post new order",
                "responses": {
                    "201": {
                        "description": "Successfully stored in database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "datas": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/order/{id}": {
            "get": {
                "tags": [
                    "order"
                ],
                "summary": "Returns specified order with details",
                "description": "Return the specified order, by ID, with details",
                "operationId": "get orders by Id",
                "parameters": [
                    {
                        "parameter": "get_order_id",
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "schema": {
                            "type": "string",
                            "default": "60b927367825c419083d3588"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully got the wanted order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nothing found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not found"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "order"
                ],
                "summary": "Update order from put form based on ID",
                "description": "Update the targeted order from form in database, using post method",
                "operationId": "Update order",
                "parameters": [
                    {
                        "parameter": "get_order_id",
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "schema": {
                            "type": "string",
                            "default": "60b927367825c419083d3588"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Succesfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully updated"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "order"
                ],
                "summary": "Delete order from Delete method based on ID",
                "description": "Delete the targeted order from form in database, using delete method",
                "operationId": "Delete order",
                "parameters": [
                    {
                        "parameter": "get_order_id",
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "schema": {
                            "type": "string",
                            "default": "60b927367825c419083d3588"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully deleted"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ID not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nothing found at this ID"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Database error"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "Current time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Order": {
                "title": "Order",
                "description": "Order Model",
                "properties": {
                    "_id": {
                        "title": "_id",
                        "description": "_id of the entity",
                        "type": "ObjectId",
                        "example": "60b927367825c419083d3588"
                    },
                    "bookingId": {
                        "title": "bookingId",
                        "type": "ObjectId",
                        "example": "60b923547825c419083d3585"
                    },
                    "services": {
                        "title": "services",
                        "type": "array",
                        "items": {}
                    },
                    "comment": {
                        "title": "comment",
                        "type": "string",
                        "example": "Intolérent au lactose"
                    },
                    "discount": {
                        "title": "discount",
                        "type": "string",
                        "example": "CouponABC123"
                    },
                    "state": {
                        "title": "discount",
                        "type": "integer",
                        "example": "1"
                    },
                    "start": {
                        "title": "start",
                        "type": "date",
                        "example": "2021-06-09 17H00"
                    },
                    "end": {
                        "title": "end",
                        "type": "date",
                        "example": "2021-06-09 19H00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Order"
                }
            },
...
        }
    },
    "security": [
        []
    ]
}